Apply eslint
