+++ src/components/Editor/CodeEditor.tsx
@@ -10,7 +10,7 @@ import randomColor from 'randomcolor';
 import ClientCursor from './ClientCursor';
 
 import { IAppState } from '../../store/store';
-import { AttachDocAction, loadDocAction } from '../../actions/docActions';
+import { AttachDoc, loadDocAction } from '../../actions/docActions';
 import { ConnectionStatus, AddPeer, DisconnectPeer } from '../../actions/peerActions';
 
 import 'codemirror/lib/codemirror.css';
@@ -34,6 +34,7 @@ export default function CodeEditor(props: CodeEditorProps) {
   const dispatch = useDispatch();
   const classes = useStyles();
 
+  const { docKey } = props;
   const doc = useSelector((state: IAppState) => state.docState.doc);
   const client = useSelector((state: IAppState) => state.docState.client);
   const loading = useSelector((state: IAppState) => state.docState.loading);
@@ -59,13 +60,13 @@ export default function CodeEditor(props: CodeEditorProps) {
   // Attach document
   useEffect(() => {
     dispatch(loadDocAction(true));
-    dispatch(AttachDocAction(props.docKey));
-  }, [props.docKey, dispatch]);
+    dispatch(AttachDoc(docKey));
+  }, [docKey, dispatch]);
 
   // Subscribe other client
   useEffect(() => {
     if (!client || !doc) {
-      return;
+      return () => {};
     }
 
     const disconnectClient = (clientId: string) => {
@@ -81,12 +82,12 @@ export default function CodeEditor(props: CodeEditorProps) {
         const newPeerClientsId: string[] = event.value[doc.getKey().toIDString()];
         const setNewPeerClientsId = new Set(newPeerClientsId);
 
-        for (const clientId of Object.keys(peerClients)) {
+        Object.keys(peerClients).forEach((clientId) => {
           if (setNewPeerClientsId.has(clientId) && peerClients[clientId].status === ConnectionStatus.Connected) {
-            continue;
+            return;
           }
           disconnectClient(clientId);
-        }
+        });
       }
     });
 
@@ -125,10 +126,10 @@ export default function CodeEditor(props: CodeEditorProps) {
           clientCursor!.updateLine(editor, fromPos, toPos);
         };
 
-        doc?.subscribe((event: any) => {
+        doc.subscribe((event: any) => {
           if (event.name === 'remote-change') {
             event.value.forEach((change: any) => {
-              const actor = change.id.actor;
+              const { actor } = change.id;
               if (actor !== client.getID()) {
                 if (!otherClientsCursor.current.has(actor)) {
                   connectClient(actor);
@@ -142,10 +143,10 @@ export default function CodeEditor(props: CodeEditorProps) {
 
         const root = doc.getRootObject() as any;
         root.content.onChanges((changes: any) => {
-          for (const change of changes) {
-            const actor = change.actor;
-            const from = change.from;
-            const to = change.to;
+          changes.forEach((change: any) => {
+            const { actor } = change;
+            const { from } = change;
+            const { to } = change;
             if (change.type === 'content') {
               const content = change.content || '';
 
@@ -170,7 +171,7 @@ export default function CodeEditor(props: CodeEditorProps) {
                 updateLine(actor, fromPos, toPos);
               }
             }
-          }
+          });
         });
 
         // We need to subtract the height of NavBar.
@@ -186,7 +187,7 @@ export default function CodeEditor(props: CodeEditorProps) {
         const from = editor.indexFromPos(data.ranges[0].anchor);
         const to = editor.indexFromPos(data.ranges[0].head);
 
-        doc?.update((root: any) => {
+        doc.update((root: any) => {
           root.content.updateSelection(from, to);
         });
       }}
@@ -199,7 +200,7 @@ export default function CodeEditor(props: CodeEditorProps) {
         const to = editor.indexFromPos(change.to);
         const content = change.text.join('\n');
 
-        doc?.update((root: any) => {
+        doc.update((root: any) => {
           root.content.edit(from, to, content);
         });
       }}
