+++ src/pages/DocPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+import { createStyles, makeStyles } from '@material-ui/core/styles';
 
 import NavBar from '../components/NavBar';
 import CodeEditor from '../components/Editor/CodeEditor';
@@ -9,7 +9,7 @@ type DocPageProps = {
   docKey: string;
 };
 
-const useStyles = makeStyles((theme: Theme) =>
+const useStyles = makeStyles(() =>
   createStyles({
     root: {
       flexGrow: 1,
@@ -19,7 +19,10 @@ const useStyles = makeStyles((theme: Theme) =>
 
 export default function DocPage(props: RouteComponentProps<DocPageProps>) {
   const classes = useStyles();
-  const docKey = props.match.params.docKey;
+  const {
+    match: { params },
+  } = props;
+  const { docKey } = params;
 
   return (
     <div className={classes.root}>
