+++ src/store/store.ts
@@ -4,17 +4,19 @@
  * combineReducers: Merges reducers into one
  * createStore: Creates a Redux store that holds the state tree
  * Store: The TS Type used for the store, or state tree
- **/
+ */
 import { applyMiddleware, combineReducers, createStore, Store } from 'redux';
 
 /**
  * Thunk
  * Redux Thunk middleware allows you to write action creators that return a function instead of an action. The thunk can be used to delay the dispatch of an action, or to dispatch only if a certain condition is met. The inner function receives the store methods dispatch and getState as parameters.
- **/
+ */
 import thunk from 'redux-thunk';
 // Import reducers and state type
-import { IDocState, docReducer } from '../reducers/docReducer';
-import { PeerState, peerReducer } from '../reducers/peerReducer';
+import { IDocState } from '../actions/docActions';
+import docReducer from '../reducers/docReducer';
+import { PeerState } from '../actions/peerActions';
+import peerReducer from '../reducers/peerReducer';
 
 // Create an interface for the application state
 export interface IAppState {
